pyproject.toml
[project]
name = "trading-analytics-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "websocket-client>=1.8.0",
]
